+++ addons/godot_state_charts/parallel_state.gd
@@ -79,29 +79,38 @@ func _state_step():
 	for child in _sub_states:
 		child._state_step()
 
-func _state_event(event:StringName) -> bool:
+func _process_transitions(event:StringName, property_change:bool = false) -> bool:
 	if not active:
 		return false
 
-	# forward event to all children
+	# forward to all children
 	var handled := false
 	for child in _sub_states:
-		var child_handled_it = child._state_event(event)
+		var child_handled_it = child._process_transitions(event, property_change)
 		handled = handled or child_handled_it
 
-	# if any child handled the event, we don't touch it anymore
+	# if any child handled this, we don't touch it anymore
 	if handled:
 		# emit the event_received signal for completeness
-		self.event_received.emit(event)
+		# unless it was a property change
+		if not property_change:
+			self.event_received.emit(event)
 		return true
 
 	# otherwise handle it ourselves
-	# base class will also emit the event_received signal
-	return super._state_event(event)
+	# defer to the base class
+	return super._process_transitions(event, property_change)
 
 func _get_configuration_warnings() -> PackedStringArray:
 	var warnings = super._get_configuration_warnings()
-	if get_child_count() == 0:
-		warnings.append("Parallel states should have at least one child state.")
+	
+	var child_count = 0
+	for child in get_children():
+		if child is State:
+			child_count += 1
+	
+	if child_count < 2:
+		warnings.append("Parallel states should have at least two child states.")
+	
 	
 	return warnings