+++ godot_state_charts_examples/automatic_transitions/stamina_bar.gd
@@ -1,0 +1,9 @@
+extends ProgressBar
+
+
+@onready var the_frog = %TheFrog
+
+
+# Called every frame. 'delta' is the elapsed time since the previous frame.
+func _process(_delta):
+	value = the_frog.stamina